+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## unreleased
+* Small styling improvement
+* Option `showSubmit` for contact forms is now named `standalone`. The old option will remain as alias until the next major version
+* Better error messages for standalone contact forms
+
 ## 1.2.0 (2019-12-03)
 * Show error when input is higher than allowed
 * Allow selecting only a few package IDs